rework test as subprocess calls
